+++ CHANGES.rst
@@ -8,6 +8,13 @@ Major changes
     columns (year, month, day, hour, minute, second, ...). It is now the default transformer used
     in the SuperVectorizer for datetime columns.
 
+Notes
+-----
+
+* The transformers_ attribute of the SuperVectorizer now contains column names
+instead of column indices for the "remainder" columns.
+
+
 Release 0.2.2
 =============
 
@@ -16,6 +23,7 @@ Bug-fixes
 
 * Fixed a bug in the :class:`SuperVectorizer` causing a `FutureWarning`
   when using the `get_feature_names_out` method.
+
 
 Release 0.2.1
 =============